# aiot-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Nunu Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/products/key/{productKey}/message-parser": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取产品唯一的消息解析脚本；尚未配置时返回默认模板。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "产品模块"
                ],
                "summary": "获取产品消息解析器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品 Key",
                        "name": "productKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-MessageParserGetProductMessageParserResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "按产品覆盖保存 JavaScript ES5 消息解析脚本。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "产品模块"
                ],
                "summary": "保存产品消息解析器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品 Key",
                        "name": "productKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MessageParserUpsertProductMessageParserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-MessageParserGetProductMessageParserResponse"
                        }
                    }
                }
            }
        },
        "/products/key/{productKey}/message-parser/execute": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "仅在受限 JavaScript ES5 运行时中执行保存的产品脚本。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "产品模块"
                ],
                "summary": "模拟执行产品消息解析器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品 Key",
                        "name": "productKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MessageParserExecuteProductMessageParserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse-MessageParserExecuteProductMessageParserResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ApiResponse-MessageParserExecuteProductMessageParserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/MessageParserExecuteProductMessageParserResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "ApiResponse-MessageParserGetProductMessageParserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/MessageParserGetProductMessageParserResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "MessageParserExecuteProductMessageParserRequest": {
            "type": "object",
            "required": [
                "mode"
            ],
            "properties": {
                "jsonObj": {
                    "type": "string"
                },
                "mode": {
                    "type": "string",
                    "enum": [
                        "device_report",
                        "device_receive",
                        "custom"
                    ]
                },
                "rawData": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "MessageParserExecuteProductMessageParserResponse": {
            "type": "object",
            "properties": {
                "jsonOutput": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "rawData": {
                    "type": "string"
                }
            }
        },
        "MessageParserGetProductMessageParserResponse": {
            "type": "object",
            "properties": {
                "parser": {
                    "$ref": "#/definitions/MessageParserProductMessageParser"
                }
            }
        },
        "MessageParserProductMessageParser": {
            "type": "object",
            "properties": {
                "isDefault": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "productKey": {
                    "type": "string"
                },
                "script": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "MessageParserUpsertProductMessageParserRequest": {
            "type": "object",
            "required": [
                "language",
                "script"
            ],
            "properties": {
                "language": {
                    "type": "string"
                },
                "script": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}