# aiot-frontend/src/api/generated/model/messageParserExecuteProductMessageParserRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */
import type { MessageParserExecuteProductMessageParserRequestMode } from './messageParserExecuteProductMessageParserRequestMode';

export interface MessageParserExecuteProductMessageParserRequest {
  jsonObj?: string;
  mode: MessageParserExecuteProductMessageParserRequestMode;
  rawData?: string;
  topic?: string;
}