# aiot-frontend/src/api/generated/model/messageParserGetProductMessageParserResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */
import type { MessageParserProductMessageParser } from './messageParserProductMessageParser';

export type MessageParserGetProductMessageParserResponse =
  MessageParserProductMessageParser;